Project01/server_py/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
PyJWT>=2.8.0
bcrypt>=4.1.0
tensorflow>=2.15.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
